+++ examples/data-sources/hcp_vault_secrets_secret/data-source.tf
@@ -1,0 +1,4 @@
+data "hcp_vault_secrets_secret" "example" {
+  app_name    = "example-vault-secrets-app"
+  secret_name = "my_secret"
+}